refactor: unify servo angle and position handling, deprecate clamping logic
- Updated servo APIs to use `float` for angle representation (0.0 - 180.0 degrees) across HAL, platform, and Python modules.
- Removed position-to-angle conversions and clamping logic for simplification.
- Adjusted default and cached servo values to reflect angle units (`float`) instead of positional integers.
- Refactored existing functions to directly operate on angles without intermediate conversions.
- Enhanced consistency in servo-related APIs and clarified angle range documentation.

+++ modules/libstp-hal/include/hal/Servo.hpp
@@ -18,7 +18,7 @@ namespace libstp::hal::servo
      */
     class Servo
     {
-        int storedPosition = 0;
+        float storedPosition = 0.0f;
 
 #ifdef SAFETY_CHECKS_ENABLED
         static inline std::set<int> used_servo_ports{};
@@ -36,10 +36,10 @@ namespace libstp::hal::servo
 
         ~Servo();
 
-        /// Command a new servo position using the active platform's units.
-        void setPosition(int position);
-        /// Return the last commanded position cached by the wrapper.
-        int getPosition() const;
+        /// Command a new servo position in degrees (0-180).
+        void setPosition(float position);
+        /// Return the last commanded position in degrees.
+        float getPosition() const;
 
         /// Enable output while preserving the last commanded position.
         void enable() const;
